+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 v0.3.0 (in development)
 -----------------------
-- Added `PropertiesFile` class for preserving comments in files
+- Added `PropertiesFile` class for preserving comments in files [#1]
 - The `ordereddict` package is now required under Python 2.6
 
 v0.2.1 (2017-03-20)